Refactor: Add *HIST.nc in Mlminimizer.py
	Move initialization of routine_properties to the Mlas
	object. Create a compute_routine_properties() method in the
	Mlas object. Create a new RoutinePropertyManager class,
	inheriting from PropertyManager. With this refactoring,
	Mlminimizer inherits the computation of routine properties.

	Move "netcdf skip" boolean (for GCMC studies) to
	initialization of *HIST.nc.
