+++ mlacs/properties/__init__.py
@@ -6,26 +6,22 @@
 // For the initials of contributors, see CONTRIBUTORS.md
 """
 
-from mlacs.properties.property_manager import PropertyManager
+from mlacs.properties.property_manager import (PropertyManager,
+                                               RoutinePropertyManager,
+                                               )
 from mlacs.properties.calc_property import (CalcPafi,
                                             CalcNeb,
                                             CalcRdf,
                                             CalcAdf,
                                             CalcTi,
                                             CalcExecFunction,
-                                            CalcRoutineFunction,
-                                            CalcPressure,
-                                            CalcProperty,
-                                            CalcAcell,
-                                            CalcAngles,
-                                            CalcSpinAt,
-                                            CalcElectronicEntropy,
                                             )
 
 from .basic_function import (eos_fit,
                              )
 
 __all__ = ['PropertyManager',
+           'RoutinePropertyManager',
            'CalcProperty',
            'CalcPafi',
            'CalcNeb',
@@ -33,11 +29,5 @@ __all__ = ['PropertyManager',
            'CalcAdf',
            'CalcTi',
            'CalcExecFunction',
-           'CalcRoutineFunction',
-           'CalcPressure',
-           'CalcAcell',
-           'CalcAngles',
-           'CalcSpinAt',
-           'CalcElectronicEntropy',
            'eos_fit',
            ]
