+++ mlacs/state/optimize_lammps_state.py
@@ -95,6 +95,7 @@ class OptimizeLammpsState(BaseLammpsState):
     >>> neb = OptimizeLammpsState(initial, pressure=0, ptype='iso')
     >>> state.run_dynamics(initial, mlip.pair_style, mlip.pair_coeff)
     """
+
     def __init__(self, min_style="cg", etol=0.0, ftol=1e-6, dt=0.5,
                  pressure=None, ptype="iso", vmax=1e-3,
                  nsteps=1000, nsteps_eq=100, logfile=None, trajfile=None,
